wasmer opcode cost file rename
