+++ vm-executor-experimental/src/middlewares.rs
@@ -1,15 +1,15 @@
 mod we_breakpoints;
 mod we_metering;
-mod we_metering_helpers;
 mod we_opcode_control;
+mod we_opcode_cost;
 mod we_opcode_trace;
 mod we_protected_globals;
 mod we_protected_globals_trait;
 
 pub(crate) use we_breakpoints::*;
 pub(crate) use we_metering::*;
-pub use we_metering_helpers::*;
 pub use we_opcode_control::OpcodeControl;
+pub use we_opcode_cost::*;
 pub use we_opcode_trace::OpcodeTracer;
 pub use we_protected_globals::ProtectedGlobals;
 pub use we_protected_globals_trait::MiddlewareWithProtectedGlobals;
