+++ vm-executor-wasmer/src/lib.rs
@@ -6,8 +6,8 @@ mod wasmer_imports;
 mod wasmer_instance;
 mod wasmer_logger;
 mod wasmer_metering;
-mod wasmer_metering_helpers;
 mod wasmer_opcode_control;
+mod wasmer_opcode_cost;
 mod wasmer_opcode_trace;
 mod wasmer_protected_globals;
 mod wasmer_service;
@@ -16,7 +16,7 @@ mod wasmer_vm_hooks;
 pub use wasmer_executor::*;
 pub use wasmer_instance::*;
 pub use wasmer_logger::*;
-pub use wasmer_metering_helpers::*;
+pub use wasmer_opcode_cost::*;
 pub use wasmer_service::*;
 
 #[cfg(feature = "multiversx-chain-vm-executor")]
